+++ app/routes/components/Markdown/changelog.codegen.ts
@@ -5,4 +5,7 @@ export const changelog = ''
   + '- 2025-10-10: Initial version.\n'
   + '- 2025-10-24: Support render code blocks (using internal rehype plugins). Wrap\n'
   + '  its children in `Typography.Root`.\n'
+  + "- 2025-11-16: Support render `strong`, `em`, `ol` elements by `Typography`'s\n"
+  + '  sub-components. Use `Typography` instead of `Typography.Root` as the root\n'
+  + '  component.\n'
   ;